+++ CATeam3LAPS/src/main/java/sg/edu/iss/LAPS/services/ManagerServiceImpl.java
@@ -142,7 +142,7 @@ public class ManagerServiceImpl implements ManagerService {
 	}
 	
 	@Override
-	public Integer increaseThisSubordinateLeaveEntitled(String mgrEmail, Long subid, Integer increaseBy) {
+	public Float increaseThisSubordinateLeaveEntitled(String mgrEmail, Long subid, Float increaseBy) {
 		LeaveEntitled subLE = lerepo.findCompensationLeaveByUserId(subid);
 		subLE.setTotalLeave(subLE.getTotalLeave() + increaseBy);
 		lerepo.saveAndFlush(subLE);
